core: force cast

+++ src/preprocess.cpp
@@ -17,7 +17,10 @@ namespace coacd
 
         for (unsigned int i = 0; i < input.points.size(); ++i)
         {
-            points.push_back({input.points[i][0] * scale, input.points[i][1] * scale, input.points[i][2] * scale});
+            points.push_back({
+                static_cast<float>(input.points[i][0] * scale), 
+                static_cast<float>(input.points[i][1] * scale), 
+                static_cast<float>(input.points[i][2] * scale)});
         }
         for (unsigned int i = 0; i < input.triangles.size(); ++i)
         {
